requirements.txt
Appium-Python-Client
fastapi[standard]